backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
PyJWT==2.8.0
django-cors-headers==4.3.1
dj-database-url==2.1.0


